+++ cm-backend/src/main/java/com/collegemates/exceptions/ResourceNotFoundException.java
@@ -10,7 +10,7 @@ public class ResourceNotFoundException extends RuntimeException {
 	String fieldName;
 	long fieldValue;
 
-	public ResourceNotFoundException(String resourceName, String fieldName, long fieldValue) { 
+	public ResourceNotFoundException(String resourceName, String fieldName, Integer fieldValue) {
 		super(String.format("%s not found with %s : %s", resourceName, fieldName, fieldValue)) ;
 		this.resourceName = resourceName;
 		this.fieldName = fieldName;
